Move new dice list to separate file
