--- src/roll/components/rolling-dice.css
@@ -1,50 +1,0 @@
-/* Cycles through dice unicode symbols. */
-
-.rolling-dice-group {
-    max-width: 100%;
-    overflow-y: auto;
-}
-
-.dice-roll-icon {
-    font-size: 1.5em;
-    font-weight: 800;
-    padding-right: 0.4em;
-}
-
-.die-icon {
-    font-size: 1.6em;
-    font-weight: 900;
-}
-
-.die-hit {
-    color: #007bff;
-}
-
-.die-glitch {
-    color: #efb107;
-}
-
-@keyframes fade-in {
-    from { opacity: 0 }
-    to { opacity: 1 }
-}
-
-/*
-
-Yes, this is a lot of CS copypasta I could do with JS or a CSS generator.
-
-*/
-
-.dice-roll-1::after {
-    content: '';
-    animation: roll-1 1s fade-in;
-}
-
-@keyframes roll-1 {
-    0%  { content: '\2685'; }
-    20% { content: '\2683'; }
-    40% { content: '\2680'; }
-    60% { content: '\2682'; }
-    80% { content: '\2681'; }
-    100% { content: '\2680'; color: 0x }
-}
